ARUHA-1462 Adds feature flag to block DB write operations
Returns 503 on every attempt to write to the DB (INSERT, UPDATE, or
DELETE statements) if the feature flag is enabled

+++ src/main/java/org/zalando/nakadi/service/StorageService.java
@@ -14,12 +14,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.TransactionException;
 import org.zalando.nakadi.domain.DefaultStorage;
 import org.zalando.nakadi.domain.Storage;
+import org.zalando.nakadi.exceptions.runtime.DbWriteOperationsBlockedException;
 import org.zalando.nakadi.exceptions.runtime.DuplicatedStorageException;
 import org.zalando.nakadi.exceptions.runtime.NoStorageException;
 import org.zalando.nakadi.exceptions.runtime.RepositoryProblemException;
 import org.zalando.nakadi.exceptions.runtime.StorageIsUsedException;
 import org.zalando.nakadi.repository.db.StorageDbRepository;
 import org.zalando.nakadi.repository.zookeeper.ZooKeeperHolder;
+import org.zalando.nakadi.util.FeatureToggleService;
 import org.zalando.problem.Problem;
 
 import javax.annotation.PostConstruct;
@@ -41,16 +43,19 @@ public class StorageService {
     private final StorageDbRepository storageDbRepository;
     private final DefaultStorage defaultStorage;
     private final CuratorFramework curator;
+    private final FeatureToggleService featureToggleService;
 
     @Autowired
     public StorageService(final ObjectMapper objectMapper,
                           final StorageDbRepository storageDbRepository,
                           @Qualifier("default_storage") final DefaultStorage defaultStorage,
-                          final ZooKeeperHolder zooKeeperHolder) {
+                          final ZooKeeperHolder zooKeeperHolder,
+                          final FeatureToggleService featureToggleService) {
         this.objectMapper = objectMapper;
         this.storageDbRepository = storageDbRepository;
         this.defaultStorage = defaultStorage;
         this.curator = zooKeeperHolder.get();
+        this.featureToggleService = featureToggleService;
     }
 
     @PostConstruct
@@ -98,6 +103,10 @@ public class StorageService {
     }
 
     public Result<Void> createStorage(final JSONObject json) {
+        if (featureToggleService.isFeatureEnabled(FeatureToggleService.Feature.DISABLE_DB_WRITE_OPERATIONS)) {
+            throw new DbWriteOperationsBlockedException("Cannot create storage: write operations on DB " +
+                    "are blocked by feature flag.");
+        }
         final String type;
         final String id;
         final JSONObject configuration;
@@ -138,6 +147,10 @@ public class StorageService {
     }
 
     public Result<Void> deleteStorage(final String id) {
+        if (featureToggleService.isFeatureEnabled(FeatureToggleService.Feature.DISABLE_DB_WRITE_OPERATIONS)) {
+            throw new DbWriteOperationsBlockedException("Cannot delete storage: write operations on DB " +
+                    "are blocked by feature flag.");
+        }
         try {
             storageDbRepository.deleteStorage(id);
         } catch (final NoStorageException e) {
